Change to use double precision, and remove implementation for k_g

+++ src/util.h
@@ -1,6 +1,6 @@
 #include <framework/mesh.h>
 #include <iostream>
 
-float findVolume(std::vector<glm::uvec3>& triangles, std::vector<Vertex>& vertices);
-float findCurvature(std::vector<glm::uvec3>& triangles, std::vector<Vertex>& vertices, std::map<int, std::vector<int>>& vertexToTri);
-float findSurfaceArea(std::vector<glm::uvec3>& triangles, std::vector<Vertex>& vertices); 
+double findVolume(std::vector<glm::uvec3>& triangles, std::vector<Vertex>& vertices);
+double findCurvature(std::vector<glm::uvec3>& triangles, std::vector<Vertex>& vertices, std::map<int, std::vector<int>>& vertexToTri);
+double findSurfaceArea(std::vector<glm::uvec3>& triangles, std::vector<Vertex>& vertices);
